requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
requests==2.31.0
python-dotenv==1.0.0

# Data processing
pandas==2.1.4
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Solana blockchain integration
solana==0.30.2

# Development tools
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Monitoring
sentry-sdk==1.39.1
prometheus-client==0.19.0
psutil==5.9.6

# Email
aiohttp==3.9.1
aiosmtplib==3.0.0
